Extracted prompt truncation function wrom work.py to utils.py so it can be used in chat_chain.py for proper truncation so its safe with usage of basic hf inference Added torch.manual_seed to basic_hf_worker.py
